jupyter/notebook/producer/requirements.txt
websockets
websocket-client
finnhub-python
kafka-python
avro